결과 오류 처리
* Game.java: 총 수익률 결과의 소수점 아래가 모두 버려지는 것을
해결하였다.

* InputError.java: 지난 주 로또 당첨 번호 입력의 중복을 검사하는 부분에서
잘못된 String 메서드를 사용하여 중복을 정확히 검사하지 못하였다. 이를
메서드를 변경하여 해결하였다.

+++ src/main/java/com/codemcd/lotto/Game.java
@@ -46,6 +46,6 @@ public class Game {
     }
 
     private double calculateEarningsRate(int money) {
-        return totalPrizeMoney / money;
+        return totalPrizeMoney / (double) money;
     }
 }
